Update TagFormatter (ts)
