+++ src/common/Utils.ts
@@ -1,6 +1,32 @@
 /// <amd-module />
 
+import { OutAttr } from "webgme/common";
 import type { GmeCore } from "./types";
+
+export function toString(attr: OutAttr): string {
+  if (attr) {
+    return attr.toString();
+  } else {
+    return "";
+  }
+}
+
+/**
+ * Given an object and a function, return a new object where the values are `fn(value)`.
+ *
+ * @param obj - The object to map over
+ * @param fn - The function mapping the `obj` values to the output values
+ * @returns A new object with the same keys but new values
+ */
+export function mapObject<T, O>(
+  obj: { [key: string]: T },
+  fn: (value: T, key: string) => O,
+): { [key: string]: O } {
+  return Object.fromEntries(
+    Object.entries(obj).map(([k, v]) => [k, fn(v, k)]),
+  );
+}
+
 export default {
   async findTaxonomyNode(
     core: GmeCore,
